[FEATURE] PyLint configuration (#6)
## 1) Description

Pylint has been added as a new linter. It has been configured, and
tested. I tweaked a little the existing `lint` command to accommodate
the addition of Pylint. Some rules have been left commented, as they
serve no purpose on the template but might be useful when used in a
project. Those rules could not have been pre-filed.

## 2) Technical choice

Various Pylint rules were disabled : 
- design
-  line-too-long
- too-many-lines
-  too-many-nested-blocks
-  typecheck
-  import-error
-  trailing-whitespace
The choice of disabling these rules is motivated by the existence of
other linters that are dedicated to some of those tasks and does a
better job at it.

Pylint output directly to the console, to avoid polluting the console
more than needed, the report has been disabled.
As checking the report once in a while might be a useful thing, I added
a TODO to find a way to add a parameter to the lint command to enable
the report. If time shows that this functionality is not needed, the
TODO will be removed.

Pylint has not been added to the pre-commit hooks, as committing bad
code is perfectly fine since we only ask for quality in the PRs, not in
each and every step of the development.