[FEATURE] Tests, coverage and benchmarks (#8)
## 1) Description

Added pytest and few plugins to help with the elaborations of tests.
A plugin for coverage has been added, as long as another for benchmarks.
The benchmark plugin work through the usage of a fixture. This brings
control over what to benchmark.

## 2) Technical choice
The coverage is done through a plugin to avoid having to handle
differences between OSs while trying to keep the console clean.
The benchmark plugin has been added to helps keeping track of
performance.