[DOCUMENTATION] Early readme (#10)
## 1) Description

Created an early version of the README file. This version is more for
internal documentation than anything else.
Added the missing coverage computation. This part was supposed to be in
another PR but didn't get committed due to pre-commit failing the
process has a hook changed a file.
Replaced the doc formatter to pydocstyle as a linter. Reworked the
existing docstrings.

## 2) Technical choice

The original formatter didn't handle the numpy, so I changed it. While
other tools were doing the same or a better work, none of them were
either maintained or stable enough to be used here. (RIP Pyment ...)