[FEATURE] Sphinx (#11)
## 1) Description

Added and configured sphinx in order to be able to generate
documentation from doc strings contained in the projects.
Added the Read The Docs theme. 
Edited the `conf.py` file to auto update according to the changes made
in the `pyproject.toml` file.

## 2) Technical choice

I didn't configure a command to run sphinx because I think it would be
better if this was done by a workflow which would create a specific PR.
This should help to avoid conflicts in auto generated files. This choice
can be discussed if needed.