[1.0] First Iteration (#13)
## 1) Description

This PR holds all the work done on this template so far.
The whole list of what has been done cannot really be putted here, but
should be documented at some point.
Please note that this template has not been yet thoroughly tested,
problems might (and will most likely do) arise. If you choose to go by
this template, feel free to give us feedback.

## 2) Technical choice

This implementation is based on python 3.10. While 3.11 is broadly
available by now, it wasn't the case when this project started.
3.11 will most likely be skipped and a better version of the template
might be released for 3.12.

## 3) Checks

- [X] My code follows the contributing guidelines
- [X] I have read the code of conduct
- [X] I have performed a self-review of my code
- [X] My changes generate no new warnings
- [X] New and existing unit tests pass locally with my changes

---------

Signed-off-by: Valentin271 <valentin.d77@free.fr>
Co-authored-by: Valentin271 <36198422+Valentin271@users.noreply.github.com>
Co-authored-by: Valentin271 <valentin.doreau@free.fr>